refactor(api): make primitive types more cohesive
refactor(types): move datatypes into package

refactor(types): remove dt.Double/dt.Float

- replace dt.Double with dt.Float64
- replace dt.Float with dt.Float32

BREAKING CHANGE: `dt.float` is now `dt.float64`; use `dt.float32` for the previous behavior.
