+++ ibis/backends/impala/tests/test_pandas_interop.py
@@ -73,7 +73,7 @@ def exhaustive_df():
                     8.8000001907348633,
                     9.8999996185302734,
                 ],
-                dtype='f4',
+                dtype='f8',
             ),
             'int_col': np.array([0, 1, 2, 3, 4, 5, 6, 7, 8, 9], dtype='i4'),
             'month': [11, 11, 11, 11, 2, 11, 11, 11, 11, 11],
