+++ ibis/backends/impala/udf.py
@@ -340,8 +340,8 @@ _impala_to_ibis_type = {
     'smallint': 'int16',
     'int': 'int32',
     'bigint': 'int64',
-    'float': 'float',
-    'double': 'double',
+    'float': 'float32',
+    'double': 'float64',
     'string': 'string',
     'varchar': 'string',
     'char': 'string',
