+++ ibis/backends/pandas/tests/execution/test_cast.py
@@ -20,8 +20,7 @@ TIMESTAMP = "2022-03-13 06:59:10.467417"
         ('float16', 'float16'),
         ('float32', 'float32'),
         ('float64', 'float64'),
-        ('double', 'float64'),
-        ('float', 'float32'),
+        ('float', 'float64'),
         ('int8', 'int8'),
         ('int16', 'int16'),
         ('int32', 'int32'),
