+++ ibis/backends/pandas/tests/test_schema.py
@@ -137,8 +137,8 @@ def test_infer_basic_types():
         ('bool_col', dt.boolean),
         ('bool_obj_col', dt.boolean),
         ('date_string_col', dt.string),
-        ('double_col', dt.double),
-        ('float_col', dt.float),
+        ('double_col', dt.float64),
+        ('float_col', dt.float32),
         ('int_col', dt.int32),
         ('month', dt.int64),
         ('smallint_col', dt.int16),
