+++ ibis/backends/pyspark/datatypes.py
@@ -27,8 +27,8 @@ _SPARK_DTYPE_TO_IBIS_DTYPE = {
     pt.BinaryType: dt.Binary,
     pt.BooleanType: dt.Boolean,
     pt.DateType: dt.Date,
-    pt.DoubleType: dt.Double,
-    pt.FloatType: dt.Float,
+    pt.DoubleType: dt.Float64,
+    pt.FloatType: dt.Float32,
     pt.ByteType: dt.Int8,
     pt.IntegerType: dt.Int32,
     pt.LongType: dt.Int64,
