+++ ibis/backends/sqlite/tests/test_functions.py
@@ -26,11 +26,11 @@ sa = pytest.importorskip("sqlalchemy")
             lambda at: sa.cast(at.c.double_col, sa.SMALLINT),
         ),
         (
-            lambda t: t.string_col.cast(dt.double),
+            lambda t: t.string_col.cast(dt.float64),
             lambda at: sa.cast(at.c.string_col, sa.REAL),
         ),
         (
-            lambda t: t.string_col.cast(dt.float),
+            lambda t: t.string_col.cast(dt.float32),
             lambda at: sa.cast(at.c.string_col, sa.REAL),
         ),
     ],
