+++ ibis/tests/expr/conftest.py
@@ -28,8 +28,8 @@ def schema():
         ('b', 'int16'),
         ('c', 'int32'),
         ('d', 'int64'),
-        ('e', 'float'),
-        ('f', 'double'),
+        ('e', 'float32'),
+        ('f', 'float64'),
         ('g', 'string'),
         ('h', 'boolean'),
         ('i', 'timestamp'),
