+++ ibis/tests/expr/mocks.py
@@ -31,8 +31,8 @@ MOCK_TABLES = {
         ('b', 'int16'),
         ('c', 'int32'),
         ('d', 'int64'),
-        ('e', 'float'),
-        ('f', 'double'),
+        ('e', 'float32'),
+        ('f', 'float64'),
         ('g', 'string'),
         ('h', 'boolean'),
         ('i', 'timestamp'),
@@ -123,8 +123,8 @@ MOCK_TABLES = {
         ('smallint_col', 'int16'),
         ('int_col', 'int32'),
         ('bigint_col', 'int64'),
-        ('float_col', 'float'),
-        ('double_col', 'double'),
+        ('float_col', 'float32'),
+        ('double_col', 'float64'),
         ('date_string_col', 'string'),
         ('string_col', 'string'),
         ('timestamp_col', 'timestamp'),
