+++ ibis/tests/expr/test_datatypes.py
@@ -184,13 +184,12 @@ def test_char_varchar_invalid(spec):
 @pytest.mark.parametrize(
     ('spec', 'expected'),
     [
-        ('any', dt.any),
-        ('null', dt.null),
         ('boolean', dt.boolean),
         ('int8', dt.int8),
         ('int16', dt.int16),
         ('int32', dt.int32),
         ('int64', dt.int64),
+        ('int', dt.int64),
         ('uint8', dt.uint8),
         ('uint16', dt.uint16),
         ('uint32', dt.uint32),
@@ -198,9 +197,7 @@ def test_char_varchar_invalid(spec):
         ('float16', dt.float16),
         ('float32', dt.float32),
         ('float64', dt.float64),
-        ('float', dt.float),
-        ('halffloat', dt.float16),
-        ('double', dt.double),
+        ('float', dt.float64),
         ('string', dt.string),
         ('binary', dt.binary),
         ('date', dt.date),
