+++ ibis/tests/sql/conftest.py
@@ -606,8 +606,8 @@ def projection_fuse_filter(con):
             ('b', 'int16'),
             ('c', 'int32'),
             ('d', 'int64'),
-            ('e', 'float'),
-            ('f', 'double'),
+            ('e', 'float32'),
+            ('f', 'float64'),
             ('g', 'string'),
             ('h', 'boolean'),
         ],
